* Fix UserList.delUser. * Receive chat broadcasts before activity is launched too. * Add scrollbar to chat window and autoscroll to bottom.
git-svn-id: http://svn.mangler.org/mangler/branches/android@864 f654edcb-a5fb-4f28-a0e7-68c327795dae

+++ project/src/org/mangler/UserList.java
@@ -82,7 +82,7 @@ public class UserList extends ListActivity {
     
 	private void delUser(short id) {
 		for(Iterator<HashMap<String, Object>> iterator = resources.iterator(); iterator.hasNext(); ) {
-			if((Integer)iterator.next().get("id") == id) {
+			if((Short)iterator.next().get("id") == id) {
 				// Remove data.
 				iterator.remove();
